refactor: quiz.js를 quiz-start와 quiz-in-game으로 분리
