+++ src/pages/game-quiz/quiz-focus-on.html
@@ -5,7 +5,7 @@
     <link rel="icon" href="/assets/icons/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta name="description" content="타이핑 실력은 물론, 상식과 개발 지식을 함께 익힐 수 있는 인터랙티브 웹사이트입니다." />
-    <title>퀴즈 - 타임 어택 모드</title>
+    <title>퀴즈 - 집중 모드</title>
     <link rel="stylesheet" href="/fonts/KIMM.css" />
     <link rel="stylesheet" href="/src/styles/common/index.css" />
     <link rel="stylesheet" href="/src/components/header/in-game-header.css" />
@@ -17,13 +17,13 @@
     <script type="module" src="/src/pages/game-quiz/quiz-in-game.js"></script>
   </head>
   <body>
-    <div class="quiz-container time-attack" data-game="quiz">
+    <div class="quiz-container focus-on" data-game="quiz">
       <h1 class="sr-only">퀴즈를 맞춰라!</h1>
-      <h2 class="sr-only">타임어택 모드</h2>
+      <h2 class="sr-only">집중 모드</h2>
       <header class="status-bar">
         <div class="score-info">
           <span>점수: 0</span>
-          <span>시간: 60초</span>
+          <span>시간: 10초</span>
         </div>
         <div class="icon-group">
           <button id="soundToggleBtn" class="icon-button">
