+++ src/pages/game-quiz/quiz-start.html
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="/fonts/KIMM.css" />
     <link rel="stylesheet" href="/src/styles/common/index.css" />
     <link rel="stylesheet" href="/src/pages/game-quiz/quiz.css" />
-    <script type="module" src="/src/pages/game-quiz/quiz.js"></script>
+    <script type="module" src="/src/pages/game-quiz/quiz-start.js"></script>
   </head>
   <body>
     <div class="quiz-container start" data-game="quiz">
