+++ src/pages/game-quiz/quiz.css
@@ -62,7 +62,8 @@
   }
 }
 
-.quiz-container.time-attack {
+.quiz-container.time-attack,
+.quiz-container.focus-on {
   display: flex;
   flex-direction: column;
   align-items: center;
